consolidate noun and verb to syntax
Since the actual code is derived for a macro anyway, there's no point in
maintaining these as separate modules.
