+++ src/parser/syntax.rs
@@ -1,5 +1,28 @@
 use initiative_macros::WordList;
 
+#[derive(Debug, PartialEq, WordList)]
+pub enum Verb {
+    Tutorial,
+    Help,
+    Quit,
+}
+
+#[cfg(test)]
+mod test_verb {
+    use super::Verb;
+
+    #[test]
+    fn from_str_test() {
+        assert_eq!(Ok(Verb::Help), "help".parse::<Verb>());
+        assert_eq!(Err(()), "potato".parse::<Verb>());
+    }
+
+    #[test]
+    fn into_string_test() {
+        assert_eq!("help", String::from(Verb::Help).as_str());
+    }
+}
+
 #[derive(Clone, Copy, Debug, PartialEq, WordList)]
 pub enum Noun {
     Building,
